dwqdqwejhhsqswsdsadwddfddffgfdcdcef

+++ app/Http/Controllers/Seller/ProductController.php
@@ -453,9 +453,13 @@ class ProductController extends Controller
                     if ($varient) {
                         if (isset($varientData['varient_image']) && is_uploaded_file($varientData['varient_image'])) {
                             $checkCount = ProductVarient::where('image', $varient->image)->count();
-                            if ($checkCount < 2) {
-                                unlink(public_path('ProductVarient/' . $varient->image));
+                            if($checkCount != 0)
+                            {
+                                if ($checkCount < 2) {
+                                    unlink(public_path('ProductVarient/' . $varient->image));
+                                }
                             }
+
     
                             $image = $varientData['varient_image'];
                             $filename = date('YmdHis') . '_' . (string) Str::uuid() . '.webp';
